refactor: separate firstLoad and jobPost controllers into different folders
